00.compare_dex_dmso/requirements.txt
appnope==0.1.4
asttokens==3.0.0
comm==0.2.3
contourpy==1.3.3
copairs==0.5.1
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
duckdb==1.3.2
executing==2.2.0
fonttools==4.59.1
ipykernel==6.30.0
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kiwisolver==1.4.9
llvmlite==0.44.0
matplotlib==3.10.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.1
packaging==25.0
pandas==2.3.2
parso==0.8.5
patsy==1.0.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
polars==1.32.3
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pycytominer==1.2.4
pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
sqlalchemy==2.0.43
stack-data==0.6.3
statsmodels==0.14.5
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
wcwidth==0.2.13
